refactor: new domains architecture
