+++ src/core/components/Filters.module.css
@@ -1,4 +1,4 @@
-@value colors: "../styles/colors.module.css";
+@value colors: "../../styles/colors.module.css";
 @value black, grey, white, orange, green, light-gray from colors;
 
 .filter {
