+++ src/core/components/Message.module.css
@@ -1,5 +1,5 @@
-@value flexUtils: "../styles/flex-utils.module.css";
-@value colors: "../styles/colors.module.css";
+@value flexUtils: "../../styles/flex-utils.module.css";
+@value colors: "../../styles/colors.module.css";
 @value gray from colors;
 
 .root {
