+++ src/core/components/Message.tsx
