+++ src/core/components/RequestName.module.css
@@ -1,4 +1,4 @@
-@value colors: "../styles/colors.module.css";
+@value colors: "../../styles/colors.module.css";
 @value text-gray, green, orange from colors;
 
 .name {
