+++ src/core/components/Table.module.css
@@ -1,4 +1,4 @@
-@value colors: "../styles/colors.module.css";
+@value colors: "../../styles/colors.module.css";
 @value table-hover, gray, light-gray from colors;
 
 .table {
