+++ src/domains/editor/_types.ts
