+++ src/domains/editor/components/EditorActions.module.css
@@ -1,4 +1,4 @@
-@value colors: "../styles/colors.module.css";
+@value colors: "../../../styles/colors.module.css";
 @value white, green, medium-blue from colors;
 
 @keyframes spin {
