+++ src/domains/editor/components/EditorActions.tsx
