+++ src/domains/editor/constants.ts
